+++ baker/algolia/configureAlgolia.ts
@@ -1,6 +1,5 @@
 import algoliasearch from "algoliasearch"
 import { Synonym } from "@algolia/client-search"
-
 import { ALGOLIA_ID } from "settings"
 import { ALGOLIA_SECRET_KEY } from "serverSettings"
 import { countries } from "clientUtils/countries"
@@ -8,7 +7,7 @@ import { countries } from "clientUtils/countries"
 // This function initializes and applies settings to the Algolia search indices
 // Algolia settings should be configured here rather than   in the Algolia dashboard UI, as then
 // they are recorded and transferrable across dev/prod instances
-export async function configureAlgolia() {
+export const configureAlgolia = async () => {
     const client = algoliasearch(ALGOLIA_ID, ALGOLIA_SECRET_KEY)
     const chartsIndex = client.initIndex("charts")
 
@@ -78,9 +77,8 @@ export async function configureAlgolia() {
 
     // Send all our country variant names to algolia as synonyms
     for (const country of countries) {
-        if (country.variantNames) {
+        if (country.variantNames)
             synonyms.push([country.name].concat(country.variantNames))
-        }
     }
 
     const algoliaSynonyms = synonyms.map((s) => {
@@ -99,6 +97,4 @@ export async function configureAlgolia() {
     })
 }
 
-if (require.main === module) {
-    configureAlgolia()
-}
+if (require.main === module) configureAlgolia()
