+++ baker/formatWordpressPost.tsx
@@ -565,3 +565,32 @@ export const formatWordpressPost = async (
         relatedCharts: post.relatedCharts,
     }
 }
+
+export const formatPost = async (
+    post: FullPost,
+    formattingOptions: FormattingOptions,
+    grapherExports?: GrapherExports
+): Promise<FormattedPost> => {
+    const html = formatLinks(post.content)
+
+    // No formatting applied, plain source HTML returned
+    if (formattingOptions.raw)
+        return {
+            ...post,
+            html,
+            footnotes: [],
+            references: [],
+            tocHeadings: [],
+            excerpt: post.excerpt || "",
+        }
+
+    // Override formattingOptions if specified in the post (as an HTML comment)
+    const options: FormattingOptions = Object.assign(
+        {
+            toc: post.type === "page",
+            footnotes: true,
+        },
+        formattingOptions
+    )
+    return formatWordpressPost(post, html, options, grapherExports)
+}
