+++ site/HeaderSearch.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { observable, action, runInAction } from "mobx"
 import { observer } from "mobx-react"
 import { SearchResults } from "./SearchResults"
-import { SiteSearchResults, siteSearch } from "site/algolia/searchClient"
+import { SiteSearchResults, siteSearch } from "./searchClient"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import { faSearch } from "@fortawesome/free-solid-svg-icons/faSearch"
